ai-service/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
spacy==3.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
